+++ AppStoreClone/View/SmallAppInfoView.swift
@@ -46,7 +46,7 @@ struct SmallAppInfoView: View {
 
                 Spacer(minLength: 0)
 
-                DownloadButton(hasPurchase: appInfo.hasInternalPurchase, isBright: false)
+                DownloadButtonStack(hasPurchase: appInfo.hasInternalPurchase, isBright: false, direction: .vertical)
             }
 
             if !index.isMultiple(of: 3) {
